Move Player settings to `menus_settings_player.cpp`
