+++ CMakeLists.txt
@@ -2545,6 +2545,7 @@ if(CLIENT)
     components/menus_settings_assets.cpp
     components/menus_settings_controls.cpp
     components/menus_settings_controls.h
+    components/menus_settings_player.cpp
     components/menus_start.cpp
     components/menus_start.h
     components/motd.cpp
